Making the things easier
After some time, I'm retaking again this project. This commit implement a new idea to make the things easier:

The CLI will receive 3 parameters: model & vendor, output type (json or plot initially) and the input file where is located the log file.

This commit simplifies the idea. For now, I'll be working with a single log file for all vendors.